feat: add `BurgerPlacesSeeder`

+++ apps/your-burger-api/src/app/burger-business/infra/persistence/burger-place/geo-point.schema.ts
@@ -5,7 +5,7 @@ export const GeoPointSchema = new EntitySchema<GeoPoint>({
   class: GeoPoint,
   tableName: 'geopoints',
   properties: {
-    id: { type: 'numeric', primary: true, autoincrement: true },
+    id: { type: 'int', primary: true, autoincrement: true },
     street: { type: 'string', unique: true },
     latitude: { type: 'decimal' },
     longitude: { type: 'decimal' },
